:hammer: Refactor pasteboard type icons
