+++ app/src/commonMain/kotlin/com/crosspaste/ui/base/IconView.kt
@@ -2,30 +2,16 @@ package com.crosspaste.ui.base
 
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import coil3.PlatformContext
-import coil3.compose.AsyncImagePainter
-import coil3.compose.SubcomposeAsyncImage
-import coil3.compose.SubcomposeAsyncImageContent
-import coil3.request.ImageRequest
-import coil3.request.crossfade
-import okio.Path
-import org.koin.compose.koinInject
 
 @Composable
 fun PasteIconButton(
@@ -65,42 +51,4 @@ fun PasteIconButton(
             }
         }
     }
-}
-
-@Composable
-fun AppImageIcon(
-    path: Path,
-    isMacStyleIcon: Boolean,
-    size: Dp = 24.dp,
-) {
-    val platformContext = koinInject<PlatformContext>()
-
-    var imageSize by remember(path) { mutableStateOf(if (isMacStyleIcon) size else (size / 24) * 20) }
-    var imagePaddingSize by remember(path) { mutableStateOf(if (isMacStyleIcon) 0.dp else (size / 24) * 2) }
-
-    SubcomposeAsyncImage(
-        modifier = Modifier.padding(imagePaddingSize).size(imageSize),
-        model =
-            ImageRequest.Builder(platformContext)
-                .data(path)
-                .crossfade(false)
-                .build(),
-        contentDescription = "Paste Icon",
-        content = {
-            when (this.painter.state.collectAsState().value) {
-                is AsyncImagePainter.State.Error -> {
-                    imageSize = (size / 24) * 20
-                    imagePaddingSize = (size / 24) * 2
-                    Icon(
-                        painter = imageSlash(),
-                        contentDescription = "Paste Icon",
-                        modifier = Modifier.padding(imagePaddingSize).size(imageSize),
-                    )
-                }
-                else -> {
-                    SubcomposeAsyncImageContent()
-                }
-            }
-        },
-    )
 }
